Make the p text-color consistent with the provided design (#788)
When working on that test **using this code** instead of working on it
on the test home (mdn learn page), the end result looks slightly
different from how it should look like.

Here is how it should look like: 
> This screen shot from the
[test](https://developer.mozilla.org/en-US/docs/Learn_web_development/Core/Structuring_content/Test_your_skills:_HTML_text_basics#task_4)
page

![Screenshot from 2025-03-11
10-10-14](https://github.com/user-attachments/assets/cd104bcf-3240-43e8-8246-9f1058ec6924)

And this is how the end result looks like when using the original code
provided:
> Look at the code at
[codePen](https://codepen.io/youssef-elatmani/pen/vEYeXqR)
> Note that I swapped *music* with *books*, in the bellow preview. That
only applied to me.

![Screenshot from 2025-03-11
10-13-13](https://github.com/user-attachments/assets/771afeeb-e553-4d20-8a56-24d51cbb5b57)